+++ sdk/storage/azure-storage-blob-cryptography/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release History
 
 ## 12.9.0-beta.1 (Unreleased)
+- Added support to set BlobParallelUploadOptions.computeMd5 so the service can perform an md5 verification.
 - Fixed a bug where the TokenCredential scope would be incorrect for custom URLs.
 - Fixed a bug where a custom application id in HttpLogOptions would not be added to the User Agent String.